+++ CHANGELOG.md
@@ -19,6 +19,8 @@ transformation model and planar and torsion angle indices;
   cost function;
 - Function `pdbanimrmch` to produce a series of Ramachandran plots for a
   multimodel PDB file;
+- Function `pdbcartinterp` to perform linear interpolation of Cartesian
+  atom coordinates between two PDB structures;
 - Option to leave only alpha carbon atoms added to `pdbbackbone`;
 - Option to consider planar angles added to `trmdistantangleindices`;
 - `trmcreate` creates an alpha carbon-based transformation model if the